EU + Waveform Testcases
This EU takes an 8-bit input command and decodes it to perform several functions. These functions briefly are: inserting a value and storing it in one of the 3 registers available, adding this value to a register or ANDing this value bit-by-bit with a register, or doing any of the previous three functions among any 2 registers of the 3. Waveform.vwf provides a sample test case that tests these 6 functionalities briefly. Waveform 4.vwf is a testbench containing all possible input combinations.